backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-httpx==0.30.0
httpx==0.27.0
ruff==0.4.4
